more 64-bit support

+++ renderlib/ImageXYZC.h
@@ -73,7 +73,7 @@ public:
   // how many channels to enable on first load by default
   static const int FIRST_N_CHANNELS = 1;
 
-  static const uint32_t IN_MEMORY_BPP = 16;
+  static const size_t IN_MEMORY_BPP = 16;
   ImageXYZC(uint32_t x,
             uint32_t y,
             uint32_t z,
